initial commit: wifi monitoring component
This component monitors the available Wifi interfaces,
and logs them into its own logfile. The following
parameters are logged:
 * association status
 * associated AP
 * link quality

Signal and noise are not logged given that there is no
standard in Linux: the drivers can either report a percentage,
or different kinds of dBm values, and can report signal without
noise.
The default deplo